using nala alert component
