--- components/ai_chat/resources/page/components/error_context_limit_reaching/style.module.scss
@@ -1,35 +1,0 @@
-// Copyright (c) 2023 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
-
-.box {
-  width: 100%;
-  background: var(--leo-color-systemfeedback-info-background);
-  border-radius: 16px;
-  padding: 16px;
-  margin-bottom: 16px;
-  display: flex;
-  gap: 16px;
-
-  h1 {
-    margin: 0;
-    padding: 0;
-    font: var(--leo-font-primary-default-semibold);
-  }
-
-  p {
-    margin: 0 0 12px 0;
-    font: var(--leo-font-primary-default-regular);
-  }
-}
-
-.icon {
-  --leo-icon-size: 18px;
-  color: var(--leo-color-systemfeedback-info-icon);
-}
-
-.actionsBox {
-  display: inline-flex;
-  gap: 10px;
-}
