--- components/ai_chat/resources/page/components/prompt_long_article/style.module.scss
@@ -1,43 +1,0 @@
-// Copyright (c) 2023 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
-
-.common {
-  h1 {
-    margin: 0;
-    padding: 0;
-    font: var(--leo-font-primary-default-semibold);
-  }
-
-  h4 {
-    font: var(--leo-font-primary-heading-h4);
-    margin: 0 0 5px 0;
-  }
-
-  p {
-    margin: 0 0 12px 0;
-    font: var(--leo-font-primary-default-regular);
-  }
-}
-
-.box {
-  width: 100%;
-  background: var(--leo-color-systemfeedback-warning-background);
-  border-radius: 16px;
-  padding: 16px;
-  margin-bottom: 16px;
-  display: flex;
-  gap: 16px;
-}
-
-.icon {
-  --leo-icon-size: 18px;
-  color: var(--leo-color-systemfeedback-warning-icon);
-}
-
-.actionsBox {
-  display: inline-flex;
-  gap: 10px;
-}
-
